refactor(vercel): extract evidence mappers into _evidence.py
Move map_vercel_deployment_status and map_vercel_deployment_logs (plus
their page-truncation helpers) out of the flat tools/__init__.py
facade into a dedicated sibling module, per review feedback that
facade files should only expose the public tool interface.
